backend/requirements.txt
# NovaPress AI v2 - Backend Dependencies
# Python 3.10+ (Windows compatible)

# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
redis>=5.2.0

# ML & Embeddings (Open Source - NO GEMINI)
sentence-transformers>=3.2.0  # BGE-M3 embeddings
# PyTorch CPU-only (no CUDA - saves ~4GB)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers>=4.46.0

# Vector Database
qdrant-client>=1.12.0

# Clustering (Windows pre-compiled wheels)
scikit-learn>=1.5.0  # Modern version with Windows wheels
# hdbscan - using sklearn.cluster.HDBSCAN instead (built-in since sklearn 1.3)
# umap-learn>=0.5.5  # Optional, comment out if compilation issues
numpy>=1.26.0

# NLP & Knowledge Graph
spacy>=3.7.0
networkx>=3.1
python-louvain>=0.16

# Web Scraping
httpx>=0.28.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
feedparser>=6.0.0
lxml>=5.0.0
lxml_html_clean>=0.1.0
playwright>=1.40.0  # Fallback for JS-rendered sites (Phase 1 - Plan Backend)

# LLM (OpenRouter)
openai>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0  # Required for pydantic EmailStr
python-dotenv>=1.0.0
bcrypt>=4.2.0
slowapi>=0.1.9  # Rate limiting

# WebSocket
websockets>=13.0

# Text-to-Speech
edge-tts>=6.1.0  # Microsoft Edge TTS for audio summaries

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1
pyyaml>=6.0.0
aiofiles>=24.1.0

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.21.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
